check

+++ app.py
@@ -1,10 +1,11 @@
-import json, os
-from config import config
+import json
 
 from flask import Flask, render_template, request, jsonify
-from flask_login import LoginManager
+from flask_cors import CORS
+from flask_login import LoginManager, login_required, current_user
 from flask_restful import Api
 
+from config import config
 from models.model import *
 from models.model_mongo import mongo
 
@@ -16,7 +17,7 @@ def create_app(env):
     if not env:
         env = 'development'
     app.config.from_object(config[env])  # from ./config.py
-
+    CORS(app)
     register_extensions(app)
 
     return app
@@ -54,46 +55,63 @@ app = create_app("production")
 
 
 @app.route('/')
-def hello():
-    # return render_template("dist/index.html")
-    return "The demo has moved to  <a href='http://ec2-54-178-101-206.ap-northeast-1.compute.amazonaws.com/' target='_blank'>AWS server</a>"
+def index():
+    return render_template("dist/index.html")
 
+
+@login_required
+@app.route('/db_status')
+def db_status():
+    pass
+    return "status"
+
+
+@login_required
 @app.route('/db_create')
 def db_create():
-    db.create_all()
-    return "Create all table!"
+    if current_user.user_id == "109269307671077250513":
+
+        db.create_all()
+        return "Create all table!"
+    else:
+        return "No authority!"
 
 
+@login_required
 @app.route('/db_drop')
 def db_drop():
-    db.drop_all()
-    return "Drop all table!"
+    if current_user.user_id == "109269307671077250513":
+        db.drop_all()
+        return "Drop all table!"
+    else:
+        return "No authority!"
 
 
-@app.route('/db_q_test')
-def db_q_test():
-    q = User.query.all()
-    result = {}
-    for qq in q:
-        result[qq.username] = qq.email
 
-    return result
-
-
-@app.route('/db_i_test', methods=["POST"])
-def db_i_test():
-    data = request.get_data()
-    data = json.loads(data)
-    print(data)
-
-    for account in data:
-        user = User(user_id=account['user_id'], username=account['user_id'], email=account['email'],
-                    account_source=account['account_source'])
-        db.session.add(user)
-
-    db.session.commit()
-
-    return jsonify(data)
+# @app.route('/db_q_test')
+# def db_q_test():
+#     q = User.query.all()
+#     result = {}
+#     for qq in q:
+#         result[qq.username] = qq.email
+#
+#     return result
+#
+#
+# @app.route('/db_i_test', methods=["POST"])
+# def db_i_test():
+#     data = request.get_data()
+#     data = json.loads(data)
+#     print(data)
+#
+#     for account in data:
+#         user = User(user_id=account['user_id'], username=account['user_id'], email=account['email'],
+#                     account_source=account['account_source'])
+#         db.session.add(user)
+#
+#     db.session.commit()
+#
+#     return jsonify(data)
 
 
 if __name__ == "__main__":
@@ -102,8 +120,8 @@ if __name__ == "__main__":
 
     @app.route('/')
     def hello():
-        # return render_template("dist/index.html")
-        return "The demo has moved to  <a href='http://ec2-54-178-101-206.ap-northeast-1.compute.amazonaws.com/' target='_blank'>AWS server</a>"
+        return render_template("/dist/index.html")
+        # return "The demo has moved to  <a href='http://ec2-54-178-101-206.ap-northeast-1.compute.amazonaws.com/' target='_blank'>AWS server</a>"
 
 
     @app.route('/db_create')
@@ -142,4 +160,6 @@ if __name__ == "__main__":
         db.session.commit()
 
         return jsonify(data)
+
+
     app.run(host="0.0.0.0", debug=True)
